Set default multiplicities to [1-1]

+++ org.eclipse.m2m.atl.atlgt.ecore2km3/src-gen/org/eclipse/m2m/km3/impl/TypedElementImpl.java
@@ -39,7 +39,7 @@ public class TypedElementImpl extends ModelElementImpl implements TypedElement {
 	 * @generated
 	 * @ordered
 	 */
-	protected static final Integer LOWER_EDEFAULT = new Integer(0);
+	protected static final Integer LOWER_EDEFAULT = new Integer(1);
 
 	/**
 	 * The cached value of the '{@link #getLower() <em>Lower</em>}' attribute.
